move EditableCmsPageCategory from DTO to QueryResult
