+++ src/Core/Domain/CmsPageCategory/QueryResult/EditableCmsPageCategory.php
@@ -24,7 +24,7 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\DTO;
+namespace PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\QueryResult;
 
 use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\Exception\CmsPageCategoryException;
 use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\ValueObject\CmsPageCategoryId;
